requirements.txt
kagglehub
pandas
numpy
opencv-python
scikit-learn
matplotlib
tensorflow>=2.10.0
seaborn
deepface
Pillow
scipy